requirements.txt
mcp>=0.1.0
praw>=7.7.0
python-dotenv>=1.0.0